feat : 게시물 수정 기능 추가 및 기존 코드 수정
- 게시물의 작성자만이 GET `/post/edit/:post_id` 를 통해 게시물
수정페이지 접근
- PATCH `/post/:post_id`를 통한 게시물 수정
- 수정 가능한 항목 : 해시태그, 별점, 게시물 내용

- 게시물 평점 => 별점으로 텍스트 수정
- 게시물 항목별 유효성체크 메소드 모듈로 분리(post-form-validation.js)
- 게시물 작성 페이지 별점 기본값 0으로 변경

+++ public/javascripts/post-detail.js
@@ -5,6 +5,7 @@ window.addEventListener('load', function() {
 	const modal = document.querySelector('#modal');
 	const addCommentButton = document.querySelector('#addComment');
 	const deleteCommentButtons = document.querySelectorAll('.deleteComment');
+	const updatePostButton = document.querySelector('#updatePost');
 	const deletePostButton = document.querySelector('#deletePost');
 
 	modal.addEventListener('click', function() {
@@ -21,9 +22,16 @@ window.addEventListener('load', function() {
 		});
 	});
 
-	deletePostButton.addEventListener('click', function() {
-		deletePostEvent();
-	});
+	if (updatePostButton && deletePostButton) {
+		updatePostButton.addEventListener('click', function() {
+			const post_id = window.location.pathname.substring(6);
+			location.href = `/post/edit/${post_id}`;
+		});
+
+		deletePostButton.addEventListener('click', function() {
+			deletePostEvent();
+		});
+	}
 });
 
 const isValidFormData = comment_body => {
